Refactor MigratorFromSqlite classes
