+++ lizmap/modules/lizmap/lib/Logger/MigratorFromSqlite.php
@@ -2,7 +2,7 @@
 
 /**
  * @author    3liz
- * @copyright 2019 3liz
+ * @copyright 2019-2026 3liz
  *
  * @see      http://3liz.com
  *
@@ -11,50 +11,10 @@
 
 namespace Lizmap\Logger;
 
-class MigratorFromSqlite
+use Lizmap\App\AbstractMigratorFromSqlite;
+
+class MigratorFromSqlite extends AbstractMigratorFromSqlite
 {
-    public function __construct() {}
-
-    public const MIGRATE_RES_OK = 1;
-    public const MIGRATE_RES_ALREADY_MIGRATED = 2;
-
-    protected function copyTable($daoSelector, $oldProfile, $newProfile, $updateSequence = true)
-    {
-        $daoNew = \jDao::get($daoSelector, $newProfile);
-        $daoSqlite = \jDao::create($daoSelector, $oldProfile);
-        $properties = array_keys($daoSqlite->getProperties());
-        foreach ($daoSqlite->findAll() as $rec) {
-            $daoRec = \jDao::createRecord($daoSelector, $newProfile);
-            foreach ($properties as $prop) {
-                $daoRec->{$prop} = $rec->{$prop};
-            }
-
-            try {
-                $daoNew->insert($daoRec);
-            } catch (\Exception $e) {
-                echo '*** Insert ERROR for the record ';
-                var_export($rec->getPk());
-                echo "\nError is: ".$e->getMessage()."\n";
-            }
-        }
-
-        if ($updateSequence) {
-            $idField = $daoNew->getProperties()[$daoNew->getPrimaryKeyNames()[0]]['fieldName'];
-            $table = $daoNew->getTables()[$daoNew->getPrimaryTable()]['realname'];
-
-            $conn = \jDb::getConnection($newProfile);
-            $rs = $conn->query('SELECT pg_get_serial_sequence('.$conn->quote($table).','.$conn->quote($idField).') as sequence_name');
-            if ($rs && ($rec = $rs->fetch())) {
-                $sequence = $rec->sequence_name;
-                if ($sequence) {
-                    $conn->query('SELECT setval('.$conn->quote($sequence).',
-                    (SELECT max('.$conn->encloseName($idField).')
-                    FROM '.$conn->encloseName($table).'))');
-                }
-            }
-        }
-    }
-
     public function migrateLog($profileName = 'lizlog', $resetBefore = false)
     {
         $profile = \jProfiles::get('jdb', $profileName);
