+++ lizmap/modules/lizmap/lib/Users/MigratorFromSqlite.php
@@ -2,7 +2,7 @@
 
 /**
  * @author    3liz
- * @copyright 2019 3liz
+ * @copyright 2019-2026 3liz
  *
  * @see      http://3liz.com
  *
@@ -11,13 +11,10 @@
 
 namespace Lizmap\Users;
 
-class MigratorFromSqlite
+use Lizmap\App\AbstractMigratorFromSqlite;
+
+class MigratorFromSqlite extends AbstractMigratorFromSqlite
 {
-    public function __construct() {}
-
-    public const MIGRATE_RES_OK = 1;
-    public const MIGRATE_RES_ALREADY_MIGRATED = 2;
-
     public function migrateUsersAndRights($resetBefore = false)
     {
         $sqliteFile = \jApp::varPath('db/jauth.db');
@@ -65,43 +62,6 @@ class MigratorFromSqlite
         return self::MIGRATE_RES_OK;
     }
 
-    protected function copyTable($daoSelector, $oldProfile, $newProfile, $updateSequence = true)
-    {
-        $daoNew = \jDao::get($daoSelector, $newProfile);
-        $daoSqlite = \jDao::create($daoSelector, $oldProfile);
-        $properties = array_keys($daoSqlite->getProperties());
-        foreach ($daoSqlite->findAll() as $rec) {
-            $daoRec = \jDao::createRecord($daoSelector, $newProfile);
-            foreach ($properties as $prop) {
-                $daoRec->{$prop} = $rec->{$prop};
-            }
-
-            try {
-                $daoNew->insert($daoRec);
-            } catch (\Exception $e) {
-                echo '*** Insert ERROR for the record ';
-                var_export($rec->getPk());
-                echo "\nError is: ".$e->getMessage()."\n";
-            }
-        }
-
-        if ($updateSequence) {
-            $idField = $daoNew->getProperties()[$daoNew->getPrimaryKeyNames()[0]]['fieldName'];
-            $table = $daoNew->getTables()[$daoNew->getPrimaryTable()]['realname'];
-
-            $conn = \jDb::getConnection($newProfile);
-            $rs = $conn->query('SELECT pg_get_serial_sequence('.$conn->quote($table).','.$conn->quote($idField).') as sequence_name');
-            if ($rs && ($rec = $rs->fetch())) {
-                $sequence = $rec->sequence_name;
-                if ($sequence) {
-                    $conn->query('SELECT setval('.$conn->quote($sequence).',
-                    (SELECT max('.$conn->encloseName($idField).')
-                    FROM '.$conn->encloseName($table).'))');
-                }
-            }
-        }
-    }
-
     protected function createUsersTables()
     {
         // retrieve the configuration of jauth
@@ -118,7 +78,7 @@ class MigratorFromSqlite
 
         $profile = \jProfiles::get('jdb', $profileName);
         if (!$profile) {
-            throw new \UnexpectedValueException("No {$profile} profile defined into profiles.ini.php", 1);
+            throw new \UnexpectedValueException("No {$profileName} profile defined into profiles.ini.php", 1);
         }
 
         if ($profile['driver'] == 'sqlite3') {
@@ -133,7 +93,7 @@ class MigratorFromSqlite
             return array($daoSelector, $profileName);
         }
 
-        // the table does not exists, let's create it
+        // the table does not exist, let's create it
         $mapper = new \jDaoDbMapper($profileName);
         $mapper->createTableFromDao($daoSelector);
